In the interest of reducing code size, moved the old CompositeModel implementations into their own module and marked them deprecated.
