+++ astropy/modeling/core.py
@@ -37,7 +37,6 @@ from ..utils import (deprecated, sharedmethod, find_current_module,
                      InheritDocstrings)
 from ..utils.codegen import make_function_with_signature
 from ..utils.exceptions import AstropyDeprecationWarning
-from ..utils.compat.odict import OrderedDict
 from .utils import (array_repr_oneline, check_broadcast, combine_labels,
                     make_binary_operator_eval, ExpressionTree,
                     IncompatibleShapeError)
@@ -45,9 +44,8 @@ from .utils import (array_repr_oneline, check_broadcast, combine_labels,
 from .parameters import Parameter, InputParameterError
 
 
-__all__ = ['Model', 'FittableModel', 'SummedCompositeModel',
-           'SerialCompositeModel', 'LabeledInput', 'Fittable1DModel',
-           'Fittable2DModel', 'custom_model', 'ModelDefinitionError']
+__all__ = ['Model', 'FittableModel', 'Fittable1DModel', 'Fittable2DModel',
+           'custom_model', 'ModelDefinitionError']
 
 
 class ModelDefinitionError(Exception):
@@ -804,6 +802,9 @@ class Model(object):
         else:
             return _prepare_outputs_model_set(self, outputs, format_info)
 
+    @deprecated('1.0',
+                alternative='Use Model operators (TODO: link to compound '
+                            'model docs')
     def add_model(self, model, mode):
         """
         Create a CompositeModel by chaining the current model with the new one
@@ -822,6 +823,9 @@ class Model(object):
         model : CompositeModel
             an instance of CompositeModel
         """
+
+        from ._compound_deprecated import (SummedCompositeModel,
+                                           SerialCompositeModel)
 
         if mode in ['parallel', 'p']:
             return SummedCompositeModel([self, model])
@@ -1873,392 +1877,3 @@ def _validate_input_shapes(inputs, argnames, n_models, model_set_axis,
                 arg_a, shape_a, arg_b, shape_b))
 
     return input_broadcast
-
-
-# *** Pre-1.0 legacy composite model classes *** #
-# TODO: These could prehaps be rewritten on top of the new composite model
-# framework, but keeping the legacy API for backwards-compatibility.
-
-class LabeledInput(OrderedDict):
-    """
-    Used by `SerialCompositeModel` and `SummedCompositeModel` to choose input
-    data using labels.
-
-    This is a container assigning labels (names) to all input data arrays to a
-    composite model.
-
-    Parameters
-    ----------
-    data : list
-        List of all input data
-    labels : list of strings
-        names matching each coordinate in data
-
-    Examples
-    --------
-    >>> y, x = np.mgrid[:5, :5]
-    >>> l = np.arange(10)
-    >>> labeled_input = LabeledInput([x, y, l], ['x', 'y', 'pixel'])
-    >>> labeled_input.x
-    array([[0, 1, 2, 3, 4],
-           [0, 1, 2, 3, 4],
-           [0, 1, 2, 3, 4],
-           [0, 1, 2, 3, 4],
-           [0, 1, 2, 3, 4]])
-    >>> labeled_input['x']
-    array([[0, 1, 2, 3, 4],
-           [0, 1, 2, 3, 4],
-           [0, 1, 2, 3, 4],
-           [0, 1, 2, 3, 4],
-           [0, 1, 2, 3, 4]])
-    """
-
-    def __init__(self, data, labels):
-        if len(labels) != len(data):
-            raise TypeError("Number of labels and data doesn't match")
-
-        super(LabeledInput, self).__init__(zip(labels, data))
-
-    def __getattr__(self, label):
-        try:
-            return self[label]
-        except KeyError:
-            raise AttributeError(label)
-
-    def __setattr__(self, label, data):
-        if label.startswith('_'):
-            super(LabeledInput, self).__setattr__(label, data)
-        else:
-            self[label] = data
-
-    def __delattr__(self, label):
-        try:
-            del self[label]
-        except KeyError:
-            raise AttributeError(label)
-
-    @property
-    def labels(self):
-        return tuple(self.keys())
-
-    def add(self, label=None, value=None, **kw):
-        """
-        Add input data to a LabeledInput object
-
-        Parameters
-        --------------
-        label : str
-            coordinate label
-        value : numerical type
-            coordinate value
-        kw : dictionary
-            if given this is a dictionary of ``{label: value}`` pairs
-        """
-
-        if ((label is None and value is not None) or
-                (label is not None and value is None)):
-            raise TypeError("Expected label and value to be defined")
-
-        kw[label] = value
-
-        self.update(kw)
-
-    def copy(self):
-        return LabeledInput(self.values(), self.labels)
-
-
-class _LabeledInputMapping(Model):
-    def __init__(self, labeled_input, inmap, outmap):
-        self._labeled_input = labeled_input
-        self._inmap = tuple(inmap)
-        self._outmap = tuple(outmap)
-        super(_LabeledInputMapping, self).__init__()
-
-    def __repr__(self):
-        return '<{0}>'.format(self.name)
-
-    @property
-    def inputs(self):
-        return self._outmap
-
-    @property
-    def outputs(self):
-        return self._inmap
-
-    @property
-    def name(self):
-        return '{0}({1} -> {2})'.format(self.__class__.__name__,
-                                        self._outmap, self._inmap)
-
-    def evaluate(self, *inputs):
-        for idx, label in enumerate(self._outmap):
-            self._labeled_input[label] = inputs[idx]
-
-        result = tuple(self._labeled_input[label] for label in self._inmap)
-
-        if len(result) == 1:
-            return result[0]
-        else:
-            return result
-
-    def __call__(self, *inputs):
-        return self.evaluate(*inputs)
-
-
-class _CompositeModel(Model):
-    _operator = None
-
-    def __init__(self, transforms, n_inputs, n_outputs, inmap=None,
-                 outmap=None):
-        """Base class for all composite models."""
-
-        self._transforms = transforms
-        param_names = []
-        for tr in self._transforms:
-            param_names.extend(tr.param_names)
-        super(_CompositeModel, self).__init__()
-        self.param_names = param_names
-        self.n_inputs = n_inputs
-        self.n_outputs = n_outputs
-        self.fittable = False
-        self._basic_transform = None
-
-        self._inmap = inmap
-        self._outmap = outmap
-
-    def __repr__(self):
-        return '<{0}([\n{1}\n])>'.format(
-            self.__class__.__name__,
-            indent(',\n'.join(repr(tr) for tr in self._transforms),
-                   width=4))
-
-    def __str__(self):
-        parts = ['Model: {0}'.format(self.__class__.__name__)]
-        for tr in self._transforms:
-            parts.append(indent(str(tr), width=4))
-        return '\n'.join(parts)
-
-    @property
-    def inputs(self):
-        return self._transforms[0].inputs
-
-    @property
-    def outputs(self):
-        return self._transforms[-1].outputs
-
-    @property
-    def n_inputs(self):
-        return self._n_inputs
-
-    @n_inputs.setter
-    def n_inputs(self, val):
-        self._n_inputs = val
-
-    @property
-    def n_outputs(self):
-        return self._n_outputs
-
-    @n_outputs.setter
-    def n_outputs(self, val):
-        self._n_outputs = val
-
-    def invert(self):
-        raise NotImplementedError("Subclasses should implement this")
-
-    @property
-    def param_sets(self):
-        all_params = tuple(m.param_sets for m in self._transforms)
-        return np.vstack(all_params)
-
-    @property
-    def parameters(self):
-        raise NotImplementedError(
-            "Composite models do not currently support the .parameters "
-            "array.")
-
-    def evaluate(self, *inputs):
-        """
-        Specialized `Model.evaluate` implementation that allows `LabeledInput`
-        inputs to be handled when calling this model.
-
-        This ignores any passed in parameter values, as _CompoundModels can't
-        be fitted anyways.
-        """
-
-        # Drop parameter arguments
-        inputs = inputs[:self.n_inputs]
-
-        if len(inputs) == 1 and isinstance(inputs[0], LabeledInput):
-            labeled_input = inputs[0].copy()
-            transform = self._make_labeled_transform(labeled_input)
-            inputs = [labeled_input[label] for label in self._inmap[0]]
-            result = transform(*inputs)
-
-            if self._transforms[-1].n_outputs == 1:
-                labeled_input[self._outmap[-1][0]] = result
-            else:
-                for label, output in zip(self._outmap[-1], result):
-                    labeled_input[label] = output
-
-            return labeled_input
-        else:
-            if self._basic_transform is None:
-                transform = self._transforms[0]
-                for t in self._transforms[1:]:
-                    transform = self._operator(transform, t)
-
-                self._basic_transform = transform
-
-            return self._basic_transform(*inputs)
-
-    def __call__(self, *inputs):
-        """
-        Specialized `Model.__call__` implementation that allows
-        `LabeledInput` inputs to be handled when calling this model.
-        """
-
-        return self.evaluate(*inputs)
-
-    def _make_labeled_transform(self, labeled_input):
-        """
-        Build up a transformation graph that incorporates the instructions
-        encoded in the `LabeledInput` object.
-
-        This requires use of the ``_inmap`` and ``_outmap`` attributes set
-        when instantiating this `_CompositeModel`.
-        """
-
-        if self._inmap is None:
-            raise TypeError("Parameter 'inmap' must be provided when "
-                            "input is a labeled object.")
-        if self._outmap is None:
-            raise TypeError("Parameter 'outmap' must be provided when "
-                            "input is a labeled object")
-
-        transforms = [self._transforms[0]]
-        previous_outmap = self._outmap[0]
-        for model, inmap, outmap in zip(self._transforms[1:], self._inmap[1:],
-                                        self._outmap[1:]):
-            mapping = _LabeledInputMapping(labeled_input, inmap,
-                                           previous_outmap)
-            transforms.append(mapping | model)
-            previous_outmap = outmap
-
-        return functools.reduce(self._operator, transforms)
-
-
-class SerialCompositeModel(_CompositeModel):
-    """
-    Composite model that evaluates models in series.
-
-    Parameters
-    ----------
-    transforms : list
-        a list of transforms in the order to be executed
-    inmap : list of lists or None
-        labels in an input instance of LabeledInput
-        if None, the number of input coordinates is exactly what
-        the transforms expect
-    outmap : list or None
-        labels in an input instance of LabeledInput
-        if None, the number of output coordinates is exactly what
-        the transforms expect
-    n_inputs : int
-        dimension of input space (e.g. 2 for a spatial model)
-    n_outputs : int
-        dimension of output
-
-    Notes
-    -----
-    Output values of one model are used as input values of another.
-    Obviously the order of the models matters.
-
-    Examples
-    --------
-    Apply a 2D rotation followed by a shift in x and y::
-
-        >>> import numpy as np
-        >>> from astropy.modeling import models, LabeledInput, SerialCompositeModel
-        >>> y, x = np.mgrid[:5, :5]
-        >>> rotation = models.Rotation2D(angle=23.5)
-        >>> offset_x = models.Shift(-4.23)
-        >>> offset_y = models.Shift(2)
-        >>> labeled_input = LabeledInput([x, y], ["x", "y"])
-        >>> transform = SerialCompositeModel([rotation, offset_x, offset_y],
-        ...                                  inmap=[['x', 'y'], ['x'], ['y']],
-        ...                                  outmap=[['x', 'y'], ['x'], ['y']])
-        >>> result = transform(labeled_input)
-    """
-
-    _operator = operator.or_
-
-    def __init__(self, transforms, inmap=None, outmap=None, n_inputs=None,
-                 n_outputs=None):
-        if n_inputs is None:
-            n_inputs = max([tr.n_inputs for tr in transforms])
-            # the output dimension is equal to the output dim of the last
-            # transform
-            n_outputs = transforms[-1].n_outputs
-        else:
-            if n_outputs is None:
-                raise TypeError("Expected n_inputs and n_outputs")
-
-        if transforms and inmap and outmap:
-            if not (len(transforms) == len(inmap) == len(outmap)):
-                raise ValueError("Expected sequences of transform, "
-                                 "inmap and outmap to have the same length")
-
-        super(SerialCompositeModel, self).__init__(
-                transforms, n_inputs, n_outputs, inmap=inmap, outmap=outmap)
-
-    @property
-    def inverse(self):
-        try:
-            transforms = []
-            for transform in self._transforms[::-1]:
-                transforms.append(transform.inverse)
-        except NotImplementedError:
-            raise NotImplementedError(
-                "An analytical inverse has not been implemented for "
-                "{0} models.".format(transform.__class__.__name__))
-        if self._inmap is not None:
-            inmap = self._inmap[::-1]
-            outmap = self._outmap[::-1]
-        else:
-            inmap = None
-            outmap = None
-        return SerialCompositeModel(transforms, inmap, outmap)
-
-
-class SummedCompositeModel(_CompositeModel):
-    """
-    Composite model that evaluates models in parallel.
-
-    Parameters
-    --------------
-    transforms : list
-        transforms to be executed in parallel
-    inmap : list or None
-        labels in an input instance of LabeledInput
-        if None, the number of input coordinates is exactly what the
-        transforms expect
-    outmap : list or None
-
-    Notes
-    -----
-    Evaluate each model separately and add the results to the input_data.
-    """
-
-    _operator = operator.add
-
-    def __init__(self, transforms, inmap=None, outmap=None):
-        n_inputs = transforms[0].n_inputs
-        n_outputs = n_inputs
-        for transform in transforms:
-            if not (transform.n_inputs == transform.n_outputs == n_inputs):
-                raise ValueError("A SummedCompositeModel expects n_inputs = "
-                                 "n_outputs for all transforms")
-
-        super(SummedCompositeModel, self).__init__(transforms, n_inputs,
-                                                   n_outputs, inmap=inmap,
-                                                   outmap=outmap)
